requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.1
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
apprise==1.8.0
argcomplete==3.4.0
asgi-lifespan==2.1.0
asgiref==3.8.1
astunparse==1.6.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
backtrader==1.9.78.123
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==3.0.0
cmdstanpy==1.2.4
colorama==0.4.6
coloredlogs==14.0
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
coolname==2.2.0
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
cycler==0.12.1
dagster==1.7.12
dagster-pipes==1.7.12
dateparser==1.2.0
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
filelock==3.15.4
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-auth==2.30.0
google-pasta==0.2.0
google-re2==1.1.20240601
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.0.3
griffe==0.47.0
grpcio==1.64.1
grpcio-health-checking==1.62.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
holidays==0.51
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.1.3
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
kubernetes==29.0.0
lazy-object-proxy==1.10.0
libclang==18.1.1
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pathspec==0.12.1
pendulum==3.0.0
pillow==10.3.0
pluggy==1.5.0
prefect==2.19.7
prison==0.2.1
prophet==1.1.5
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
QuantLib==1.34
readchar==4.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.14.0
setproctitle==1.3.3
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stanio==0.5.0
structlog==24.2.0
sympy==1.12.1
TA-Lib==0.4.31
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.14.1
toml==0.10.2
tomli==2.0.1
toposort==1.10
torch==2.3.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.2.3
wheel==0.43.0
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
xgboost==2.1.0
yarl==1.9.4
zipp==3.19.2
